requirements.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.1
Brlapi==0.8.2
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.2
contourpy==1.2.1
cupshelpers==1.0
cycler==0.12.1
dasbus==1.4
datasets==2.19.1
dbus-python==1.2.18
decorator==4.4.2
defog==0.65.7
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
einops==0.8.0
ethtool==0.15
eval_type_backport==0.2.0
exceptiongroup==1.2.1
fastapi==0.110.2
file-magic==0.4.0
filelock==3.13.4
flash-attn==2.5.6
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gpg==1.15.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
idna==2.10
importlib_resources==6.4.0
iniparse==0.4
interegular==0.3.3
Jinja2==3.1.3
joblib==1.4.0
jsonpointer==2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lark==1.1.9
libcomps==0.1.18
llvmlite==0.42.0
lm-format-enforcer==0.9.8
lxml==4.6.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mistralai==0.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
nftables==0.1
ninja==1.11.1.1
numba==0.59.1
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
outlines==0.0.34
packaging==20.9
pandas==2.2.2
pcp==5.0
peft==0.11.0
perf==0.1
pexpect==4.8.0
pillow==10.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.8.0
psycopg2-binary==2.9.9
ptyprocess==0.6.0
pwinput==1.0.3
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycairo==1.20.1
pycups==2.0.1
pycurl==7.43.0.6
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyGObject==3.40.1
pyinotify==0.9.6
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-linux-procfs==0.7.0
pytz==2024.1
pyudev==0.22.0
PyYAML==5.4.1
ray==2.11.0
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rpm==4.16.1.3
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
selinux==3.5
sentencepiece==0.2.0
sepolicy==3.5
setools==4.4.1
setroubleshoot==3.3.31
shtab==1.7.1
six==1.15.0
sniffio==1.3.1
sos==4.5.1
sqlcoder==0.2.5
sqlparse==0.5.1
SSSDConfig==2.8.2
starlette==0.37.2
subscription-manager==1.29.33.1
sympy==1.12
systemd-python==234
texttable==1.7.0
tiktoken==0.6.0
tokenizers==0.19.1
tqdm==4.66.2
transformers==4.40.1
triton==2.2.0
trl==0.8.6
typing_extensions==4.11.0
tyro==0.8.4
tzdata==2024.1
urllib3==1.26.5
uvicorn==0.29.0
uvloop==0.19.0
vllm==0.4.1
vllm-nccl-cu12==2.18.1.0.4.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
xformers==0.0.25
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.18.2
